refactor(sdk): restructure exports into per-export folders with barrels
