+++ packages/sdk/src/core/types.ts
@@ -214,4 +214,3 @@ export type TracesConfig = CommonConfig & {
 export interface WebSdk {
   shutdown(): Promise<void>;
 }
-export type WebSdkFactory<T> = (config?: T) => WebSdk;
