+++ packages/sdk/src/index.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export type {
+  LogsConfig,
+  RootConfig,
+  TracesConfig,
+  WebSdk,
+} from './core/types.ts';
+export type { QuickStartConfig } from './startBrowserSdk.ts';
+export { quickStartBrowserSdk, startBrowserSdk } from './startBrowserSdk.ts';
