+++ packages/sdk/src/logs/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export type { LogsConfig, WebSdk } from '../core/types.ts';
+export { startLogsSdk } from './startLogsSdk.ts';
