+++ packages/sdk/src/traces/index.ts
@@ -1,0 +1,7 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export type { TracesConfig, WebSdk } from '../core/types.ts';
+export { startTracesSdk } from './startTracesSdk.ts';
