+++ packages/sdk/tsdown.config.ts
@@ -4,9 +4,9 @@ import baseConfig from '../../tsdown.config.ts';
 export default defineConfig({
   ...baseConfig,
   entry: {
-    logs: 'src/initializer/logs.ts',
-    traces: 'src/initializer/traces.ts',
-    start: 'src/initializer/start.ts',
+    index: 'src/index.ts',
+    'logs/index': 'src/logs/index.ts',
+    'traces/index': 'src/traces/index.ts',
     'session/index': 'src/session/index.ts',
   },
 });
